# backend/docs/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WatchReleaseCreate": {
        "example": {
          "artist": "Aphex Twin",
          "currency": "USD",
          "discogs_master_id": 6211,
          "discogs_release_id": 249393,
          "is_active": true,
          "match_mode": "master_release",
          "min_condition": "VG+",
          "target_price": 35,
          "title": "Selected Ambient Works 85-92",
          "year": 1992
        },
        "properties": {
          "artist": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "currency": {
            "default": "USD",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          },
          "discogs_master_id": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discogs Master Id"
          },
          "discogs_release_id": {
            "minimum": 1.0,
            "title": "Discogs Release Id",
            "type": "integer"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "match_mode": {
            "default": "exact_release",
            "enum": [
              "exact_release",
              "master_release"
            ],
            "title": "Match Mode",
            "type": "string"
          },
          "min_condition": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Condition"
          },
          "target_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Price"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "maximum": 9999.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "discogs_release_id",
          "title"
        ],
        "title": "WatchReleaseCreate",
        "type": "object"
      },
      "WatchReleaseOut": {
        "example": {
          "artist": "Aphex Twin",
          "created_at": "2026-01-20T11:52:00+00:00",
          "currency": "USD",
          "discogs_master_id": 6211,
          "discogs_release_id": 249393,
          "id": "6608868d-f10f-4a5e-a38a-2e9fa5063f85",
          "is_active": true,
          "match_mode": "master_release",
          "min_condition": "VG+",
          "target_price": 35,
          "title": "Selected Ambient Works 85-92",
          "updated_at": "2026-01-20T12:00:00+00:00",
          "user_id": "8f2a5009-c0a2-4f90-8f1b-c1716c26bf06",
          "year": 1992
        },
        "properties": {
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "discogs_master_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discogs Master Id"
          },
          "discogs_release_id": {
            "title": "Discogs Release Id",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "match_mode": {
            "enum": [
              "exact_release",
              "master_release"
            ],
            "title": "Match Mode",
            "type": "string"
          },
          "min_condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Condition"
          },
          "target_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Price"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "user_id",
          "discogs_release_id",
          "discogs_master_id",
          "match_mode",
          "title",
          "artist",
          "year",
          "target_price",
          "currency",
          "min_condition",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "WatchReleaseOut",
        "type": "object"
      },
      "WatchReleaseUpdate": {
        "example": {
          "currency": "USD",
          "is_active": true,
          "min_condition": "NM",
          "target_price": 30
        },
        "properties": {
          "artist": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "currency": {
            "anyOf": [
              {
                "maxLength": 3,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "discogs_master_id": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discogs Master Id"
          },
          "discogs_release_id": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discogs Release Id"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "match_mode": {
            "anyOf": [
              {
                "enum": [
                  "exact_release",
                  "master_release"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Mode"
          },
          "min_condition": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Condition"
          },
          "target_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Price"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 300,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "year": {
            "anyOf": [
              {
                "maximum": 9999.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "title": "WatchReleaseUpdate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "watch-wax-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/watch-releases": {
      "get": {
        "operationId": "list_watch_releases_api_watch_releases_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchReleaseOut"
                  },
                  "title": "Response List Watch Releases Api Watch Releases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Watch Releases",
        "tags": [
          "watch-releases"
        ]
      },
      "post": {
        "operationId": "create_watch_release_api_watch_releases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchReleaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchReleaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Watch Release",
        "tags": [
          "watch-releases"
        ]
      }
    },
    "/api/watch-releases/{watch_release_id}": {
      "delete": {
        "operationId": "disable_watch_release_api_watch_releases__watch_release_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "watch_release_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Watch Release Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchReleaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Disable Watch Release",
        "tags": [
          "watch-releases"
        ]
      },
      "get": {
        "operationId": "get_watch_release_api_watch_releases__watch_release_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "watch_release_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Watch Release Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchReleaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Watch Release",
        "tags": [
          "watch-releases"
        ]
      },
      "patch": {
        "operationId": "update_watch_release_api_watch_releases__watch_release_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "watch_release_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Watch Release Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchReleaseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchReleaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Watch Release",
        "tags": [
          "watch-releases"
        ]
      }
    },
...
  }
}